Refactor other components out of app.js
