+++ my-app/src/App.js
@@ -2,6 +2,8 @@ import React, { useEffect, useState } from 'react';
 import { BrowserRouter as Router, Route, Switch, Redirect, useParams } from 'react-router-dom';
 import './App.css';
 import CountyDetail from './components/CountyDetail';
+import SearchBar from './components/SearchBar';
+import CountyCardList from './components/CountyCardList';
 import { Toolbar } from '@material-ui/core';
 
 function App() {
@@ -25,8 +27,10 @@ function App() {
   );
 }
 
+/*
+  Component representing the full Home Page
+*/
 function HomePage() {
-
   return (
     <main className="home-page">
       <SearchBar/>
@@ -36,20 +40,10 @@ function HomePage() {
   );
 }
 
-function SearchBar() {
 
-  const [input, setInput] = useState("");
-
-  return (
-    <div className="page-header">
-      <h2>COVID-19 Cases Today Across the Country</h2>
-      <p>BaseCheck wants to ensure that every person has the accessible opportunity to stay well-informed about the pandemic.</p>
-      <input type="text" value={input} placeHolder="search for a county" onChange={e => setInput(e.target.value)} className="search"></input>
-      <a href={'/search/' + input} className="search-button">Search!</a>
-    </div>
-  );
-}
-
+/*
+  Component representing the search page
+*/
 function SearchPage() {
 
   let { county } = useParams();
@@ -76,40 +70,6 @@ function SearchPage() {
     </main>
   );
 
-}
-
-/*
-  Component that will act a a container to hold each CountyCard Component.
-*/
-function CountyCardList(props) {
-  console.log(props.counties);
-
-  if (props.counties) {
-    if (props.counties.length > 0) {
-      let counties = props.counties.map((county) => {
-        return <div><a href={"/county/" + county.county + "/" + county.province}>{county.county + ", " + county.province}</a></div>
-      });
-      return (
-        <div className="list">
-          <p>{props.counties.length} search results found for "{props.search}"</p>
-        <div className="card-container">
-          {counties}
-        </div>
-      </div>
-      );
-    } else {
-      return (
-        <div className="empty-list">
-          <p>Unfortunately, we have no results for "{props.search}"</p>
-          <img src="../error.jpg" alt="no results found"></img>
-        </div>
-      );
-    }
-  } else  {
-    return (
-      <div></div>
-    );
-  }
 }
 
 export default App;
